pyproject.toml
[project]
name = "cli-v1-0"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.14.0",
    "mutagen>=1.47.0",
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
    "yt-dlp>=2026.3.17",
]
